+++ Source/Processors/RecordNode/OpenEphysFormat/OriginalRecording.cpp
@@ -445,7 +445,7 @@ void OriginalRecording::writeTTLEvent(int eventIndex, const MidiMessage& event)
 	diskWriteLock.exit();
 }
 
-void OriginalRecording::writeSynchronizedData(int writeChannel, int realChannel, const float* dataBuffer, const float* ftsBuffer, int size) {};
+void OriginalRecording::writeSynchronizedData(int writeChannel, int realChannel, const float* dataBuffer, const double* ftsBuffer, int size) {};
 
 void OriginalRecording::writeData(int writeChannel, int realChannel, const float* buffer, int size)
 {
